Properties

Properties

sol_properties(prop)

Arguments

prop
string: if provided, return only the property matching this name

Value

data.frame

See also

sol_set_property

Examples

sol_properties() ## all properties that solong knows about

#> # A tibble: 28 x 4
#>    property                  units  class_name  
#>    <chr>                     <chr>  <chr>       
#>  1 lower rostral length      mm     solprop_LRL 
#>  2 lower hood length         mm     solprop_LHL 
#>  3 hood length               mm     solprop_HL  
#>  4 crest length              mm     solprop_CL  
#>  5 mantle length             mm     solprop_ML  
#>  6 shell length              mm     solprop_SHL 
#>  7 standard length           mm     solprop_SL  
#>  8 carapace length           mm     solprop_CPL 
#>  9 removed carapace length   mm     solprop_RCPL
#> 10 total length              mm     solprop_TL  
#> 11 fork length               mm     solprop_FL  
#> 12 dry weight                g      solprop_DW  
#> 13 wet weight                g      solprop_WW  
#> 14 carbon weight             g      solprop_CW  
#> 15 lipid weight              g      solprop_LpW 
#> 16 otolith length            mm     solprop_OL  
#> 17 otolith width             mm     solprop_OW  
#> 18 oral-atrial length        mm     solprop_OAL 
#> 19 ash-free dry weight       g      solprop_AFDW
#> 20 energy density dry weight kJ g-1 solprop_EDDW
#> 21 energy density wet weight kJ g-1 solprop_EDWW
#> 22 lipid content dry weight  %      solprop_LpDW
#> 23 lipid content wet weight  %      solprop_LpWW
#> 24 shell-free dry weight     g      solprop_SFDW
#> 25 water content wet weight  %      solprop_WCWW
#> 26 total energy content      kJ     solprop_TEC 
#> 27 oxygen consumption rate   ug h-1 solprop_OCR 
#> 28 age                       yr     solprop_age 
#>    notes                                                                        
#>    <chr>                                                                        
#>  1 ""                                                                           
#>  2 ""                                                                           
#>  3 ""                                                                           
#>  4 ""                                                                           
#>  5 ""                                                                           
#>  6 ""                                                                           
#>  7 ""                                                                           
#>  8 ""                                                                           
#>  9 ""                                                                           
#> 10 Note that the definition of total length may be species-specific, e.g. for A~
#> 11 ""                                                                           
#> 12 ""                                                                           
#> 13 ""                                                                           
#> 14 ""                                                                           
#> 15 ""                                                                           
#> 16 ""                                                                           
#> 17 ""                                                                           
#> 18 Used for body length of salps, see e.g. Foxton P (1966) The distribution and~
#> 19 ""                                                                           
#> 20 ""                                                                           
#> 21 ""                                                                           
#> 22 ""                                                                           
#> 23 ""                                                                           
#> 24 ""                                                                           
#> 25 ""                                                                           
#> 26 ""                                                                           
#> 27 ""                                                                           
#> 28 ""