Von Bertalanffy growth equation with optional propagation of uncertainty

Intended for use inside of sol_equation objects.

sol_vonbert(
  t,
  Linf,
  k,
  t0,
  Linf_se,
  k_se,
  t0_se,
  reps = 1000L,
  ci = 0.95,
  method = "monte carlo"
)

Arguments

t
numeric: prediction times

Linf
numeric: Linf parameter estimate

k
numeric: k parameter estimate

t0
numeric: t0 parameter estimate

Linf_se
numeric: Linf parameter standard error estimate

k_se
numeric: k parameter standard error estimate

t0_se
numeric: t0 parameter standard error estimate

reps
integer: number of Monte-Carlo samples to draw

ci
numeric: confidence level

method
string: method to use for uncertainty propagation (only "monte carlo" supported at the moment)

Value

A tibble with columns allometric_value, allometric_value_lower, allometric_value_upper. If any of the standard error inputs are missing, NULL, or NA, the upper and lower estimates will be NA_real_

Examples

sol_vonbert(0:7, Linf = 80.7, Linf_se = 0.82, k = 0.25, k_se = 0.01, t0 = -2.31, t0_se = 0.01)

#> # A tibble: 8 x 3
#>   allometric_value allometric_value_lower allometric_value_upper
#>              <dbl>                  <dbl>                  <dbl>
#> 1             35.4                   33.3                   37.5
#> 2             45.4                   43.0                   47.8
#> 3             53.2                   50.7                   55.7
#> 4             59.3                   56.9                   61.7
#> 5             64.0                   61.7                   66.4
#> 6             67.7                   65.4                   69.9
#> 7             70.6                   68.4                   72.7
#> 8             72.8                   70.7                   74.8