Allometric equation data.

A dataset containing allometric equations (relationships between the size of an organism and the size of its body parts).

allometric_equations

Format

A data frame with variables:

equation_id

the identifier of this equation

taxon_name

name of the taxon

taxon_aphia_id

the Aphia ID of the taxon (identifier within the World Register of Marine Species)

equation

a function encoding the allometric equation

inputs

a data.frame specifying (in order) the inputs to the equation, with columns 'property' (the property name of the measurement needed, e.g. 'lower rostral length'), and 'units' (the units of the measurement needed, e.g. 'mm')

return_property

the name of the body size characteristic that is estimated by this equation (e.g. 'mantle length')

return_units

the units of measurement of the returned property

reliability

a data.frame with indicators of the reliability of the equation: type (a description of how the reliability was assessed, e.g. 'R^2' or 'N' the sample size used by the authors of the equation) and value (its value)

notes

notes

reference

the source of the equation